Fixed the save function
Browser doesn't let the data url open in a new tab so I found an approach online to download it instead. However it looks a little strange because the
image shows with a transparent background - when i drew just with black ink and downloaded it looked like there was nothing there.